sls update
