+++ infer-models/.serverless/requirements
@@ -1,0 +1,1 @@
+/Users/jossai/Library/Caches/serverless-python-requirements/e3126fcbdcf8f19994f7da2bb2ce6d75714889ca4f5273b1e9fcd3d4e55e8117_x86_64_slspyc
