+++ infer-models/package.json
@@ -1,0 +1,5 @@
+{
+  "devDependencies": {
+    "serverless-python-requirements": "^6.1.0"
+  }
+}
