+++ infer-models/requirements.txt
@@ -1,0 +1,4 @@
+botocore 
+boto3 
+langchain-community 
+Pillow
