+++ streamlit_app/app.py
@@ -195,7 +195,6 @@ if submit_button2:
     st.success('Generated image deleted successfully.')
 
 # Example Prompts Section
-st.write("## Model Prompts")
 
 # Model prompts structured for Streamlit display
 model_prompts = [
@@ -210,7 +209,7 @@ model_prompts = [
     {
         'title': 'Meta Models',
         'prompts': [
-            {"Model ID": "meta.llama3-8b-instruct-v1:0", "Prompt": "Do sentiment analysis with nuances in reasoning of the following: Recently, I purchased a new smartphone from a well-known electronics brand, and I'm thoroughly disappointed. The phone frequently freezes, the battery life is terrible, and the customer service was unhelpful when I reached out for support. I expected much better quality and service based on their reputation.", "Usecase": "Sentiment analysis"},
+            {"Model ID": "meta.llama3-8b-instruct-v1:0", "Prompt": "Do sentiment analysis with nuances in reasoning of the following: Recently, I purchased a new smartphone from a well-known electronics brand, and I'm thoroughly disappointed. The phone frequently freezes, the battery life is terrible, and the customer service was unhelpful when I reached out for support. I expected much better quality and service based on their reputation.", "Usecase": "Sentiment analysis "},
             {"Model ID": "meta.llama3-70b-instruct-v1:0", "Prompt": "Provide me text classification of the following: Global warming is a major environmental issue that is causing the earth's temperatures to rise. This leads to more severe weather patterns, including increased hurricanes, droughts, and floods. Various international organizations are stepping up efforts to reduce carbon emissions and promote green energy solutions to combat climate change.", "Usecase": "Text classification"},
 
             {"Model ID": "meta.llama2-13b-chat-v1", "Prompt": "Solve the following problem: I went to the market and bought 10 apples. I gave 2 to a friend and 2 to a helper, bought 5 more, ate 1. How many apples do I have left?", "Usecase": "Math"},
