+++ streamlit_app/invoke_agent.py
@@ -20,8 +20,8 @@ from PIL import Image
 #echo $AWS_SESSION_TOKEN
 
 
-agentId = "xx" #INPUT YOUR AGENT ID HERE
-agentAliasId = "xx" # Hits draft alias, set to a specific alias id for a deployed version
+agentId = "PBNU1ZUAEK" #INPUT YOUR AGENT ID HERE
+agentAliasId = "E4AZJMRRYN" # Hits draft alias, set to a specific alias id for a deployed version
 bucket_name = 'bedrock-agent-images'
 os.environ["AWS_REGION"] = "us-west-2"
