--- x86_64/Dockerfile.python3.11
@@ -1,19 +1,0 @@
-FROM scratch
-
-ADD 29080084a5bc69d83b9a7092af379c4705455dcfa189b99a0e03543848a3bbcd.tar.xz /
-ADD 2c4f5fd4f5365007f1808a27468387a0cb0c6acd2ae1374b2d482f7ae53356b2.tar.xz /
-ADD 32f0ef17450ac33f2c6e357e7f41d5bf208ef097965b85e425f7bac241ecba3f.tar.xz /
-ADD 8322f567d46193925b9059da311a42864c19330612ca156474063573ce04a9b1.tar.xz /
-ADD 8cd02b2a421e7e6d8927434ac5b4baf46683ccf2b2c38237b75e1abfe895e063.tar.xz /
-ADD feb815887f22ce6ea2814b716320e5a1b255f00f3ed05a1f21f0d8640b117208.tar.xz /
-
-ENV LANG=en_US.UTF-8
-ENV TZ=:/etc/localtime
-ENV PATH=/var/lang/bin:/usr/local/bin:/usr/bin/:/bin:/opt/bin
-ENV LD_LIBRARY_PATH=/var/lang/lib:/lib64:/usr/lib64:/var/runtime:/var/runtime/lib:/var/task:/var/task/lib:/opt/lib
-ENV LAMBDA_TASK_ROOT=/var/task
-ENV LAMBDA_RUNTIME_DIR=/var/runtime
-
-WORKDIR /var/task
-
-ENTRYPOINT ["/lambda-entrypoint.sh"]
